DEV: Update tests
Mostly qunit-dom usage.
